code refactoring:  added development configuration
