+++ src/main/java/sia/tacocloud/config/DevelopmentConfig.java
@@ -1,0 +1,43 @@
+package sia.tacocloud.config;
+
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import sia.tacocloud.data.IngredientRepository;
+import sia.tacocloud.data.UserRepository;
+import sia.tacocloud.entity.Ingredient;
+import sia.tacocloud.entity.User;
+
+@Profile("!prod")
+@Configuration
+public class DevelopmentConfig {
+    @Bean
+    public CommandLineRunner dataLoader(IngredientRepository repo) {
+        return args -> {
+            repo.deleteAll(); // TODO: Quick hack to avoid tests from stepping on each other with constraint violations
+            repo.save(new Ingredient("FLTO", "Flour Tortilla", Ingredient.Type.WRAP));
+            repo.save(new Ingredient("COTO", "Corn Tortilla", Ingredient.Type.WRAP));
+            repo.save(new Ingredient("GRBF", "Ground Beef", Ingredient.Type.PROTEIN));
+            repo.save(new Ingredient("CARN", "Carnitas", Ingredient.Type.PROTEIN));
+            repo.save(new Ingredient("TMTO", "Diced Tomatoes", Ingredient.Type.VEGGIES));
+            repo.save(new Ingredient("LETC", "Lettuce", Ingredient.Type.VEGGIES));
+            repo.save(new Ingredient("CHED", "Cheddar", Ingredient.Type.CHEESE));
+            repo.save(new Ingredient("JACK", "Monterrey Jack", Ingredient.Type.CHEESE));
+            repo.save(new Ingredient("SLSA", "Salsa", Ingredient.Type.SAUCE));
+            repo.save(new Ingredient("SRCR", "Sour Cream", Ingredient.Type.SAUCE));
+        };
+    }
+
+    @Bean
+    public CommandLineRunner defaultUserDataLoader(UserRepository userRepo) {
+        BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+        return args -> {
+            userRepo.deleteAll(); // TODO: Quick hack to avoid tests from stepping on each other with constraint violations
+            userRepo.save(new User(
+                    "user", passwordEncoder.encode("ppp"),
+                    "fullname", "street", "city", "CA", "zip", "123456789"));
+        };
+    }
+}
